make all tests async, copy data during migration
