+++ .isort.cfg
@@ -2,4 +2,4 @@
 line_length = 88
 multi_line_output = 3
 include_trailing_comma = True
-known_third_party = alembic,asyncpg,boto3,botocore,click,docker,fastapi,fiona,geoalchemy2,geojson,gino,gino_starlette,httpx,moto,pendulum,psycopg2,pydantic,pyproj,pytest,rasterio,requests,shapely,sqlalchemy,sqlalchemy_utils,starlette
+known_third_party = alembic,asyncpg,boto3,botocore,click,docker,fastapi,fiona,geoalchemy2,geojson,gino,gino_starlette,httpx,mock,moto,pendulum,psycopg2,pydantic,pyproj,pytest,rasterio,requests,shapely,sqlalchemy,sqlalchemy_utils,starlette
