+++ tests/tasks/test_default_assets.py
@@ -10,22 +10,16 @@ from app.application import ContextEngine, db
 from app.crud import assets, tasks, versions
 from app.models.enum.assets import AssetStatus, AssetType
 from app.models.orm.geostore import Geostore
-from app.routes.tasks.tasks import (
-    _register_dynamic_vector_tile_cache,
-    _update_asset_field_metadata,
-)
-from app.tasks.default_assets import create_default_asset
 from app.utils.aws import get_s3_client
 
-from .. import BUCKET, GEOJSON_NAME, SHP_NAME, TSV_NAME, TSV_PATH
-from . import check_callbacks, create_dataset, create_version, poll_jobs
+from .. import BUCKET, GEOJSON_NAME, PORT, SHP_NAME, TSV_NAME, TSV_PATH
+from ..utils import create_default_asset
 
 
 @pytest.mark.asyncio
-async def test_vector_source_asset(batch_client, httpd):
+async def test_vector_source_asset(batch_client, async_client):
 
     _, logs = batch_client
-    httpd_port = httpd.server_port
 
     ############################
     # Setup test
@@ -34,40 +28,34 @@ async def test_vector_source_asset(batch_client, httpd):
     dataset = "test"
     sources = (SHP_NAME, GEOJSON_NAME)
 
-    await create_dataset(dataset)
-
     for i, source in enumerate(sources):
         version = f"v1.1.{i}"
         input_data = {
-            "source_type": "vector",
-            "source_uri": [f"s3://{BUCKET}/{source}"],
             "creation_options": {
-                "src_driver": "GeoJSON",
+                "source_type": "vector",
+                "source_uri": [f"s3://{BUCKET}/{source}"],
+                "source_driver": "GeoJSON",
                 "zipped": False,
                 "create_dynamic_vector_tile_cache": True,
             },
             "metadata": {},
         }
-        await create_version(dataset, version, input_data)
 
-        ######################
-        # Test asset creation
-        #####################
-        # Create default asset in mocked BATCH
-        async with ContextEngine("WRITE"):
-            asset_id = await create_default_asset(dataset, version, input_data, None)
-
-        tasks_rows = await tasks.get_tasks(asset_id)
-        task_ids = [str(task.task_id) for task in tasks_rows]
-
-        # make sure, all jobs completed
-        status = await poll_jobs(task_ids)
-
-        # Get the logs in case something went wrong
-        _print_logs(logs)
-        check_callbacks(task_ids, httpd_port)
-
-        assert status == "saved"
+        # we only need to create the dataset once
+        if i > 0:
+            skip_dataset = True
+        else:
+            skip_dataset = False
+        asset = await create_default_asset(
+            dataset,
+            version,
+            version_payload=input_data,
+            async_client=async_client,
+            logs=logs,
+            execute_batch_jobs=True,
+            skip_dataset=skip_dataset,
+        )
+        asset_id = asset["asset_id"]
 
         await _check_version_status(dataset, version)
         await _check_asset_status(dataset, version, 1)
@@ -89,23 +77,22 @@ async def test_vector_source_asset(batch_client, httpd):
 
         await _check_dynamic_vector_tile_cache_status(dataset, version)
 
-        requests.delete(f"http://localhost:{httpd.server_port}")
+        requests.delete(f"http://localhost:{PORT}")
 
 
 @pytest.mark.asyncio
-async def test_table_source_asset(batch_client, httpd):
+async def test_table_source_asset(batch_client, async_client):
 
     _, logs = batch_client
-    httpd_port = httpd.server_port
 
     ############################
     # Setup test
     ############################
-
-    s3_client = get_s3_client()
-
-    s3_client.create_bucket(Bucket=BUCKET)
-    s3_client.upload_file(TSV_PATH, BUCKET, TSV_NAME)
+    #
+    # s3_client = get_s3_client()
+    #
+    # s3_client.create_bucket(Bucket=BUCKET)
+    # s3_client.upload_file(TSV_PATH, BUCKET, TSV_NAME)
 
     dataset = "table_test"
     version = "v202002.1"
@@ -128,10 +115,10 @@ async def test_table_source_asset(batch_client, httpd):
                 pass
 
     input_data = {
-        "source_type": "table",
-        "source_uri": [f"s3://{BUCKET}/{TSV_NAME}"],
         "creation_options": {
-            "src_driver": "text",
+            "source_type": "table",
+            "source_uri": [f"s3://{BUCKET}/{TSV_NAME}"],
+            "source_driver": "text",
             "delimiter": "\t",
             "has_header": True,
             "latitude": "latitude",
@@ -162,28 +149,19 @@ async def test_table_source_asset(batch_client, httpd):
         "metadata": {},
     }
 
-    await create_dataset(dataset)
-    await create_version(dataset, version, input_data)
-
     #####################
     # Test asset creation
     #####################
 
-    # Create default asset in mocked BATCH
-    async with ContextEngine("WRITE"):
-        asset_id = await create_default_asset(dataset, version, input_data, None,)
-
-    tasks_rows = await tasks.get_tasks(asset_id)
-    task_ids = [str(task.task_id) for task in tasks_rows]
-
-    # make sure, all jobs completed
-    status = await poll_jobs(task_ids)
-
-    # Get the logs in case something went wrong
-    _print_logs(logs)
-    check_callbacks(task_ids, httpd_port)
-
-    assert status == "saved"
+    asset = await create_default_asset(
+        dataset,
+        version,
+        version_payload=input_data,
+        execute_batch_jobs=True,
+        logs=logs,
+        async_client=async_client,
+    )
+    asset_id = asset["asset_id"]
 
     await _check_version_status(dataset, version)
     await _check_asset_status(dataset, version, 1)
@@ -234,18 +212,17 @@ async def test_table_source_asset(batch_client, httpd):
 
 
 @pytest.mark.asyncio
-async def test_table_source_asset_parallel(batch_client, httpd):
+async def test_table_source_asset_parallel(batch_client, async_client):
     _, logs = batch_client
-    httpd_port = httpd.server_port
 
     ############################
     # Setup test
     ############################
-
-    s3_client = get_s3_client()
-
-    s3_client.create_bucket(Bucket=BUCKET)
-    s3_client.upload_file(TSV_PATH, BUCKET, TSV_NAME)
+    #
+    # s3_client = get_s3_client()
+    #
+    # s3_client.create_bucket(Bucket=BUCKET)
+    # s3_client.upload_file(TSV_PATH, BUCKET, TSV_NAME)
 
     dataset = "table_test"
     version = "v202002.1"
@@ -268,11 +245,11 @@ async def test_table_source_asset_parallel(batch_client, httpd):
                 pass
 
     input_data = {
-        "source_type": "table",
-        "source_uri": [f"s3://{BUCKET}/{TSV_NAME}"]
-        + [f"s3://{BUCKET}/test_{i}.tsv" for i in range(2, 101)],
         "creation_options": {
-            "src_driver": "text",
+            "source_type": "table",
+            "source_uri": [f"s3://{BUCKET}/{TSV_NAME}"]
+            + [f"s3://{BUCKET}/test_{i}.tsv" for i in range(2, 101)],
+            "source_driver": "text",
             "delimiter": "\t",
             "has_header": True,
             "latitude": "latitude",
@@ -303,28 +280,19 @@ async def test_table_source_asset_parallel(batch_client, httpd):
         "metadata": {},
     }
 
-    await create_dataset(dataset)
-    await create_version(dataset, version, input_data)
-
     #####################
     # Test asset creation
     #####################
 
-    # Create default asset in mocked BATCH
-    async with ContextEngine("WRITE"):
-        asset_id = await create_default_asset(dataset, version, input_data, None,)
-
-    tasks_rows = await tasks.get_tasks(asset_id)
-    task_ids = [str(task.task_id) for task in tasks_rows]
-
-    # make sure, all jobs completed
-    status = await poll_jobs(task_ids)
-
-    # Get the logs in case something went wrong
-    _print_logs(logs)
-    check_callbacks(task_ids, httpd_port)
-
-    assert status == "saved"
+    asset = await create_default_asset(
+        dataset,
+        version,
+        version_payload=input_data,
+        execute_batch_jobs=True,
+        logs=logs,
+        async_client=async_client,
+    )
+    asset_id = asset["asset_id"]
 
     await _check_version_status(dataset, version)
     await _check_asset_status(dataset, version, 1)
@@ -392,44 +360,29 @@ def _assert_fields(field_list, field_schema):
     assert count == len(field_schema)
 
 
-def _print_logs(logs):
-    resp = logs.describe_log_streams(logGroupName="/aws/batch/job")
-
-    for stream in resp["logStreams"]:
-        ls_name = stream["logStreamName"]
-
-        stream_resp = logs.get_log_events(
-            logGroupName="/aws/batch/job", logStreamName=ls_name
-        )
-
-        print(f"-------- LOGS FROM {ls_name} --------")
-        for event in stream_resp["events"]:
-            print(event["message"])
-
-
 async def _check_version_status(dataset, version):
     row = await versions.get_version(dataset, version)
 
     # in this test we don't set the final version status to saved or failed
-    assert row.status == "pending"
+    assert row.status == "saved"
 
     # in this test we only see the logs from background task, not from batch jobs
     print(f"TABLE SOURCE VERSION LOGS: {row.change_log}")
-    assert len(row.change_log) == 1
+    assert len(row.change_log) == 3
     assert row.change_log[0]["message"] == "Successfully scheduled batch jobs"
 
 
 async def _check_asset_status(dataset, version, nb_assets):
     rows = await assets.get_assets(dataset, version)
-    assert len(rows) == 1
+    assert len(rows) == 2
 
     # in this test we don't set the final asset status to saved or failed
-    assert rows[0].status == "pending"
+    assert rows[0].status == "saved"
     assert rows[0].is_default is True
 
     # in this test we only see the logs from background task, not from batch jobs
     print(f"TABLE SOURCE ASSET LOGS: {rows[0].change_log}")
-    assert len(rows[0].change_log) == nb_assets
+    assert len(rows[0].change_log) == nb_assets * 2
 
 
 async def _check_task_status(asset_id, nb_jobs, last_job_name):
@@ -446,31 +399,18 @@ async def _check_task_status(asset_id, nb_jobs, last_job_name):
 async def _check_dynamic_vector_tile_cache_status(dataset, version):
     rows = await assets.get_assets(dataset, version)
     asset_row = rows[0]
-    asset_row = await _update_asset_field_metadata(
-        asset_row.dataset, asset_row.version, asset_row.asset_id,
-    )
 
     # SHP files have one additional attribute (fid)
     if asset_row.version == "v1.1.0":
-        assert len(asset_row.metadata["fields"]) == 10
+        assert len(asset_row.fields) == 10
     else:
-        assert len(asset_row.metadata["fields"]) == 9
-
-    # We need the asset status in saved to create dynamic vector tile cache
-    async with ContextEngine("WRITE"):
-        asset_row = await assets.update_asset(
-            asset_row.asset_id, status=AssetStatus.saved
-        )
-
-    await _register_dynamic_vector_tile_cache(
-        asset_row.dataset, asset_row.version, asset_row.metadata
-    )
+        assert len(asset_row.fields) == 9
 
     rows = await assets.get_assets(dataset, version)
     v = await versions.get_version(dataset, version)
     print(v.change_log)
 
     assert len(rows) == 2
-    assert rows[0].asset_type == AssetType.database_table
+    assert rows[0].asset_type == AssetType.geo_database_table
     assert rows[1].asset_type == AssetType.dynamic_vector_tile_cache
     assert rows[1].status == AssetStatus.saved
